feat(security,frontend,db): Hardening, Vue 2.7.16 upgrade, and database index optimization (#0.96.0)
- Security & Backdoor Removal:
  - Removed GO_ENV=test backdoor in handlers/authentication.go for 100% environment parity.
  - Resolved IsUser() API key and Bearer token privilege resolution on private service queries.
  - Added live HTTP security regression test suite enforcing 401 unauthenticated responses and secret redaction.

- Database & Schema Indexing Optimization:
  - Added composite indexes (idx_hits_service_created_at, idx_failures_service_created_at, idx_checkin_hits_created_at) for time-series query acceleration.
  - Added indexed api_key lookup on users table.
  - Enforced foreign key constraints (ON DELETE CASCADE) across relational tables.

- Frontend & Supply Chain Hardening:
  - Upgraded Vue to Vue 2.7.16 (final 2.x release), Axios to ^0.33.0, and Sass to ^1.101.7.
  - Upgraded Vue ecosystem: vue-router (^3.6.5), vuex (^3.6.2), vue-i18n (^8.28.2), vuedraggable (^2.24.3), vue-cookies (^1.8.4), core-js (^3.38.1), date-fns (^2.30.0), codemirror (^5.65.18), fontawesome (^5.15.4).
  - Upgraded devDependencies: @babel/core & @babel/preset-env (^7.26.0), @vue/cli-service (^4.5.19), @vue/test-utils (^1.3.6), clean-webpack-plugin (^4.0.0), webpack-bundle-analyzer (^4.10.2), stylelint (^13.13.1).
  - Configured NPM overrides (lodash, qs, serialize-javascript, elliptic, express, node-forge) and committed package-lock.json.
  - Created frontend/.yarnrc (ignore-engines true) for Node 24 build compatibility.
  - Fixed JS/Vue diagnostics across 84 application source files (oxlint & biome 0 errors).

- Tooling & Code Hygiene:
  - Cleaned all shell scripts for 100% ShellCheck compliance.
  - Created code_quality.sh running all 8 quality and security audit phases in sequence.

+++ install.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
-# Statping installation script for Linux, Mac, and maybe Windows.
-#
-# This installation script is a modification of Yarn's installation
+# Statping installation script for Linux, Mac, and Windows.
 set -e
 
 reset="\033[0m"
@@ -10,40 +8,38 @@ green="\033[32m"
 yellow="\033[33m"
 cyan="\033[36m"
 white="\033[37m"
-gpg_key=64B9C6AAE2D55278
-gpgurl=https://statping.com/statping.gpg
-repo=https://github.com/statping-ng/statping-ng
+repo="https://github.com/statping-ng/statping-ng"
 
 statping_get_tarball() {
+  local os_type="$1"
+  local arch_type="$2"
   fext='tar.gz'
-  if [ ${OS} == 'windows' ]; then
+  if [ "${os_type}" = 'windows' ]; then
     fext='zip'
-    ARCH='x64'
+    arch_type='x64'
   fi
-  url="$repo/releases/latest/download/statping-$1-$2.$fext"
-  printf "$cyan> Downloading latest version for $OS $ARCH...\n$url $reset\n"
-  # Get both the tarball and its GPG signature
-  tarball_tmp=`mktemp -t statping.tar.gz.XXXXXXXXXX`
+  url="${repo}/releases/latest/download/statping-${os_type}-${arch_type}.${fext}"
+  printf "%b> Downloading latest version for %s %s...\n%s %b\n" "$cyan" "$os_type" "$arch_type" "$url" "$reset"
+  tarball_tmp=$(mktemp -t statping.tar.gz.XXXXXXXXXX)
   if curl --fail -L -s -o "$tarball_tmp" "$url"; then
-    # All this dance is because `tar --strip=1` does not work everywhere
     temp=$(mktemp -d statping.XXXXXXXXXX)
-    if [ ${OS} == 'windows' ]; then
-      unzip $tarball_tmp -d "$temp"
+    if [ "${os_type}" = 'windows' ]; then
+      unzip "$tarball_tmp" -d "$temp"
     else
-      tar xzf $tarball_tmp -C "$temp"
+      tar xzf "$tarball_tmp" -C "$temp"
     fi
-    printf "$green> Installing to $DEST/statping-ng\n"
+    printf "%b> Installing to %s/statping-ng\n" "$green" "$DEST"
     mv "$temp"/statping "$DEST"
     rm -rf "$temp"
-    rm $tarball_tmp*
-    printf "$cyan> Statping-ng is now installed! $reset\n"
-    printf "$white>   Repo:     $repo $reset\n"
-    printf "$white>   Wiki:     $repo/wiki $reset\n"
-    printf "$white>   Issues:   $repo/issues $reset\n"
-    printf "$cyan> Try to run \"statping help\" $reset\n"
+    rm -f "${tarball_tmp:?}"*
+    printf "%b> Statping-ng is now installed! %b\n" "$cyan" "$reset"
+    printf "%b>   Repo:     %s %b\n" "$white" "$repo" "$reset"
+    printf "%b>   Wiki:     %s/wiki %b\n" "$white" "$repo" "$reset"
+    printf "%b>   Issues:   %s/issues %b\n" "$white" "$repo" "$reset"
+    printf "%b> Try to run \"statping help\" %b\n" "$cyan" "$reset"
   else
-    printf "$red> Failed to download $url.$reset\n"
-    exit 1;
+    printf "%b> Failed to download %s.%b\n" "$red" "$url" "$reset"
+    exit 1
   fi
 }
 
@@ -52,96 +48,79 @@ statping_reset() {
 }
 
 statping_brew_install() {
-  if [[ -z "$(command -v brew --version)" ]]; then
-    printf "${white}Using Brew to install!$reset\n"
-    printf "${yellow}---> brew tap statping/statping$reset\n"
+  if command -v brew > /dev/null 2>&1; then
+    printf "%bUsing Brew to install!%b\n" "$white" "$reset"
+    printf "%b---> brew tap statping-ng/statping-ng%b\n" "$yellow" "$reset"
     brew tap statping-ng/statping-ng
-    printf "${yellow}---> brew install statping$reset\n"
+    printf "%b---> brew install statping%b\n" "$yellow" "$reset"
     brew install statping-ng
-    printf "${green}Brew installation is complete!$reset\n"
-    printf "${yellow}You can use 'brew upgrade' to upgrade Statping next time.$reset\n"
+    printf "%bBrew installation is complete!%b\n" "$green" "$reset"
+    printf "%bYou can use 'brew upgrade' to upgrade Statping next time.%b\n" "$yellow" "$reset"
   else
-    statping_get_tarball $OS $ARCH
+    statping_get_tarball "$OS" "$ARCH"
   fi
 }
 
 statping_install() {
-  printf "${white}Installing Statping-ng!$reset\n"
+  printf "%bInstalling Statping-ng!%b\n" "$white" "$reset"
   getOS
   getArch
-  statping_get_tarball $OS $ARCH
+  statping_get_tarball "$OS" "$ARCH"
   statping_reset
 }
 
 statping_verify_or_quit() {
-  read -p "$1 [y/N] " -n 1 -r
+  read -r -p "$1 [y/N] " REPLY
   echo
-  if [[ ! $REPLY =~ ^[Yy]$ ]]
-  then
-    printf "$red> Aborting$reset\n"
+  if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    printf "%b> Aborting%b\n" "$red" "$reset"
     exit 1
   fi
 }
 
-# get the users operating system
 getOS() {
-    OS="`uname`"
-    case $OS in
-      'Linux')
-        OS='linux'
-        DEST=/usr/local/bin
-        alias ls='ls --color=auto'
-        ;;
-      'FreeBSD')
-        OS='freebsd'
-        DEST=/usr/local/bin
-        alias ls='ls -G'
-        ;;
-      'OpenBSD')
-        OS='openbsd'
-        DEST=/usr/local/bin
-        alias ls='ls -G'
-        ;;
-      'WindowsNT')
-        OS='windows'
-        DEST=/usr/local/bin
-        ;;
-      'MINGW*')
-        OS='windows'
-        DEST=/usr/local/bin
-        ;;
-      'CYGWIN*')
-        OS='windows'
-        DEST=/usr/local/bin
-        ;;
-      'Darwin')
-        OS='darwin'
-        DEST=/usr/local/bin
-        ;;
-      'SunOS')
-        OS='linux'
-        DEST=/usr/local/bin
-        ;;
-      'AIX') ;;
-      *) ;;
-    esac
+  OS="$(uname)"
+  case $OS in
+    'Linux')
+      OS='linux'
+      DEST=/usr/local/bin
+      ;;
+    'FreeBSD')
+      OS='freebsd'
+      DEST=/usr/local/bin
+      ;;
+    'OpenBSD')
+      OS='openbsd'
+      DEST=/usr/local/bin
+      ;;
+    'WindowsNT' | 'MINGW*' | 'CYGWIN*')
+      OS='windows'
+      DEST=/usr/local/bin
+      ;;
+    'Darwin')
+      OS='darwin'
+      DEST=/usr/local/bin
+      ;;
+    'SunOS')
+      OS='linux'
+      DEST=/usr/local/bin
+      ;;
+    *) ;;
+  esac
 }
 
-# get 64x or 32 machine arch
 getArch() {
-    MACHINE_TYPE=`uname -m`
-    if [ ${MACHINE_TYPE} == 'x86_64' ]; then
-      ARCH="amd64"
-    elif [ ${MACHINE_TYPE} == 'arm' ]; then
-      ARCH="arm"
-    elif [ ${MACHINE_TYPE} == 'arm64' ]; then
-      ARCH="arm64"
-    elif [ ${MACHINE_TYPE} == 'aarch64' ]; then
-      ARCH="arm64"
-    else
-      ARCH="386"
-    fi
+  MACHINE_TYPE=$(uname -m)
+  if [ "${MACHINE_TYPE}" = 'x86_64' ]; then
+    ARCH="amd64"
+  elif [ "${MACHINE_TYPE}" = 'arm' ]; then
+    ARCH="arm"
+  elif [ "${MACHINE_TYPE}" = 'arm64' ] || [ "${MACHINE_TYPE}" = 'aarch64' ]; then
+    ARCH="arm64"
+  else
+    ARCH="386"
+  fi
 }
 
-cd ~
-statping_install $1 $2
+cd ~ || exit 1
+statping_install "$1" "$2"
